added the storm manager class
