+++ Assets/Prefabs/Worlds/World6.prefab
@@ -83,6 +83,7 @@ GameObject:
   m_Component:
   - component: {fileID: 2048080934235256897}
   - component: {fileID: 5199821468771705673}
+  - component: {fileID: 965882122287875450}
   m_Layer: 0
   m_Name: World6
   m_TagString: Untagged
@@ -122,6 +123,18 @@ MonoBehaviour:
   m_EditorClassIdentifier: 
   _rotationDuration: 0.2
   _rotableChild: {fileID: 3909547111023414187}
+--- !u!114 &965882122287875450
+MonoBehaviour:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 2159126921825663166}
+  m_Enabled: 1
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 11500000, guid: ed3fea2907bf8564685924048d3bb868, type: 3}
+  m_Name: 
+  m_EditorClassIdentifier: 
   _stormEffect: {fileID: 2100000, guid: 8251dd813068e6a408526df7a699b7bb, type: 2}
   _alphaCurve:
     serializedVersion: 2
@@ -129,29 +142,20 @@ MonoBehaviour:
     - serializedVersion: 3
       time: 0
       value: 0
-      inSlope: 2.1449852
-      outSlope: 2.1449852
+      inSlope: 2
+      outSlope: 2
       tangentMode: 0
       weightedMode: 0
       inWeight: 0
-      outWeight: 0.15127727
-    - serializedVersion: 3
-      time: 0.5233216
-      value: 0.86334294
-      inSlope: 0.7111872
-      outSlope: 0.7111872
-      tangentMode: 0
-      weightedMode: 0
-      inWeight: 0.33333334
-      outWeight: 0.20083502
+      outWeight: 0
     - serializedVersion: 3
       time: 1
       value: 1
-      inSlope: 0.023103638
-      outSlope: 0.023103638
+      inSlope: 0
+      outSlope: 0
       tangentMode: 0
       weightedMode: 0
-      inWeight: 0.42831114
+      inWeight: 0
       outWeight: 0
     m_PreInfinity: 2
     m_PostInfinity: 2
